move external mappings from model to subsytem class
