--- numerous/engine/model/external_mappings/__init__.py
@@ -1,2 +1,0 @@
-from .external_mappings import *
-from .interpolation_type import *
