+++ numerous/engine/system/external_mappings/interpolation_type.py
